rename Table folder to make it easier to find
