+++ src/components/index.scss
@@ -2,5 +2,5 @@
 @import "./Depiction/depiction";
 @import "./Icon/icon";
 @import "./Tree/tree";
-@import "./SimpleTable/simpletable";
+@import "./Table/table";
 @import "./Typography/typography";
