+++ src/components/index.ts
@@ -2,7 +2,7 @@ export * from "./Badge/Badge";
 export * from "./Depiction/Depiction";
 export * from "./Icon";
 export * from "./ProgressBar";
-export * from "./SimpleTable";
+export * from "./Table";
 export * from "./Tag";
 export * from "./Tree/Tree";
 export * from "./Typography";
